xllite: bugfix: change rt._batch type from int to uint32_t to fix compilation attention "comparison of integer expressions of different signedness"

+++ xlite/csrc/runtime.cpp
@@ -448,7 +448,7 @@ void XRuntime::PrepareAttn(XModelAttnMeta &attnMeta, uint64_t maxBatchedTokens, 
 
     batchedTokens = 0;
     _maxNumBlocks = 0;
-    _batch = static_cast<int>(batch);
+    _batch = batch;
     queryStart = 0;
     bool allCached = true;
     bool anyMultiToken = false;
